tests/inputdevice-touch.html

<!DOCTYPE html>
<title>InputDevice test for touch-triggered events</title>
<meta name="viewport" content="width=device-width">
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevice-tests.js"></script>
<style>
#target {
  border: 1px solid black;
  width: 100px;
  height: 100px;
  background-color: lightgreen;
  margin-bottom: 20px;
}
</style>

<h1>InputDevice - test for touch-triggered events</h1>
<p>Tap and then drag on the touchscreen in the box below, then tap outside the box.</p>
<div id="target" tabindex=1></div>
<div><b>Expected events:</b> <span id="eventList"></span></div>
<div id="log"></div>
<script>
'use strict';

function inputDeviceValidator(sourceDevice) {
  assert_true(event.sourceDevice != null, "sourceDevice is set");
  assert_true(event.sourceDevice.firesTouchEvents, "sourcedevice.firesTouchEvents");
}

function runTests() {
  var events = ["touchstart", "touchmove", "touchend", "mousedown", "mousemove", "mouseup", "mouseenter", "mouseleave", "mouseover", "mouseout", "click", "focus", "focusin", "focusout", "blur"];
  if ('PointerEvent' in window)
    events = events.join(["pointerdown", "pointermove", "pointerup", "pointerenter", "pointerleave", "pointerover", "pointerout", "pointercancel"]);
  runInputDeviceEventTests(events, inputDeviceValidator);
}
</script>
